refactor(evals): split memory suites 6-7 into their own helper modules
Lift runAgentDay into memory-suite-agent-day.ts and runLatencySuite into
memory-suite-latency.ts so memory-suites.ts stays under 500 LOC; the shared
supersede() helper moves to memory-harness.ts. Pure move, no behavior change.
